requirements.txt
flax
numpy
pillow
pyyaml
absl-py